refactor: Move kotlin-library plugin to own module
