+++ infrastructure-android/src/main/kotlin/AndroidLibraryPlugin.kt
@@ -2,7 +2,7 @@ package com.redmadrobot.build
 
 import com.android.build.gradle.LibraryExtension
 import com.redmadrobot.build.internal.android
-import com.redmadrobot.build.internal.kotlin
+import com.redmadrobot.build.kotlin.internal.kotlin
 import org.gradle.api.Project
 import org.jetbrains.kotlin.gradle.dsl.ExplicitApiMode
 import org.jetbrains.kotlin.gradle.tasks.KotlinCompile
