+++ infrastructure-android/src/main/kotlin/BaseAndroidPlugin.kt
@@ -3,8 +3,8 @@ package com.redmadrobot.build
 import com.android.build.gradle.BaseExtension
 import com.redmadrobot.build.extension.AndroidOptions
 import com.redmadrobot.build.internal.android
-import com.redmadrobot.build.internal.configureKotlin
-import com.redmadrobot.build.internal.setTestOptions
+import com.redmadrobot.build.kotlin.internal.configureKotlin
+import com.redmadrobot.build.kotlin.internal.setTestOptions
 import org.gradle.api.JavaVersion
 import org.gradle.api.Project
 import org.gradle.api.provider.Property
