+++ infrastructure-base/build.gradle.kts
@@ -1,9 +1,1 @@
-gradlePlugin {
-    plugins {
-        isAutomatedPublishing = false
-        register("root-project") {
-            id = "redmadrobot.root-project"
-            implementationClass = "com.redmadrobot.build.RootProjectPlugin"
-        }
-    }
-}
+
