+++ infrastructure-kotlin/src/main/kotlin/kotlin/internal/Kotlin.kt
@@ -1,4 +1,4 @@
-package com.redmadrobot.build.internal
+package com.redmadrobot.build.kotlin.internal
 
 import com.redmadrobot.build.InfrastructurePlugin
 import com.redmadrobot.build.dsl.isRunningOnCi
