+++ infrastructure/build.gradle.kts
@@ -3,9 +3,9 @@ description = "Small plugins to reduce boilerplate in Gradle build scripts."
 gradlePlugin {
     plugins {
         isAutomatedPublishing = false
-        register("kotlin-library") {
-            id = "redmadrobot.kotlin-library"
-            implementationClass = "com.redmadrobot.build.KotlinLibraryPlugin"
+        register("root-project") {
+            id = "redmadrobot.root-project"
+            implementationClass = "com.redmadrobot.build.RootProjectPlugin"
         }
     }
 }
@@ -16,9 +16,7 @@ repositories {
 }
 
 dependencies {
-    api(projects.infrastructureBase)
-    api(kotlin("gradle-plugin", version = libs.versions.kotlin.get()))
-    compileOnly(libs.androidGradle)
-    implementation(libs.detektGradle)
-    implementation(libs.jgit)
+    api(projects.infrastructurePublish)
+    api(projects.infrastructureDetekt)
+    api(projects.infrastructureKotlin)
 }
